Create Stock Price Action.py
A deep model that was responsible for predicting the price of stocks with the historical data of their respective prices.  This Long Short-Term Memory (LSTM) Network being the improved version of RNN was built using keras, sklearn, pandas, etc. The DL model takes sequential data of a particular stock selected and gets complied with loss as MSE, it provides a result of the predicted output alongside the actual output in a single graph.